[giac] initial changes for Numworks N110 port changes for xcas to cpp translator
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@69330 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ giac/src/giac/cpp/prog.cc
@@ -9102,7 +9102,7 @@ namespace giac {
     return RTC_GetTicks();
 #else
 #if defined __APPLE__ || defined EMCC || !defined HAVE_LIBRT
-    return clock();
+    return (int) clock();
 #else
     timespec t;
     clock_gettime(CLOCK_MONOTONIC,&t);
